tools/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
socksio>=1.0.0